Avoid split of functionality across files
