+++ cpp/dolfinx/fem/CMakeLists.txt
@@ -16,7 +16,6 @@ set(HEADERS_fem
   ${CMAKE_CURRENT_SOURCE_DIR}/discreteoperators.h
   ${CMAKE_CURRENT_SOURCE_DIR}/dofmapbuilder.h
   ${CMAKE_CURRENT_SOURCE_DIR}/dolfin_fem.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/evaluate.h
   ${CMAKE_CURRENT_SOURCE_DIR}/interpolate.h
   ${CMAKE_CURRENT_SOURCE_DIR}/petsc.h
   ${CMAKE_CURRENT_SOURCE_DIR}/sparsitybuild.h
